+++ docs/changelog.rst
@@ -11,6 +11,20 @@ Version 0.10.0
 Not yet released
 
 
+Version 0.9.2
+==============
+
+Not yet released
+
+
+Fixes
+-----
+
+* Revert changes in :class:`aspecd.utils.ToDictMixin` from version 0.9.1, as it caused problems with Matplotlib.
+* Updates on contour plots to work with Matplotlib 3.8
+* Updates to prevent deprecation warning for NumPy 1.25
+
+
 Version 0.9.1
 ==============